+++ tools/scylla-sstable.cc
@@ -32,6 +32,7 @@
 #include "types/map.hh"
 #include "tools/json_writer.hh"
 #include "tools/schema_loader.hh"
+#include "tools/sstable_consumer.hh"
 #include "tools/utils.hh"
 #include "locator/host_id.hh"
 
@@ -157,31 +158,6 @@ const std::vector<sstables::shared_sstable> load_sstables(schema_ptr schema, sst
 
     return sstables;
 }
-
-// stop_iteration::no -> continue consuming sstable content
-class sstable_consumer {
-public:
-    virtual ~sstable_consumer() = default;
-    // called at the very start
-    virtual future<> on_start_of_stream() = 0;
-    // stop_iteration::yes -> on_end_of_sstable() - skip sstable content
-    // sstable parameter is nullptr when merging multiple sstables
-    virtual future<stop_iteration> on_new_sstable(const sstables::sstable* const) = 0;
-    // stop_iteration::yes -> consume(partition_end) - skip partition content
-    virtual future<stop_iteration> consume(partition_start&&) = 0;
-    // stop_iteration::yes -> consume(partition_end) - skip remaining partition content
-    virtual future<stop_iteration> consume(static_row&&) = 0;
-    // stop_iteration::yes -> consume(partition_end) - skip remaining partition content
-    virtual future<stop_iteration> consume(clustering_row&&) = 0;
-    // stop_iteration::yes -> consume(partition_end) - skip remaining partition content
-    virtual future<stop_iteration> consume(range_tombstone_change&&) = 0;
-    // stop_iteration::yes -> on_end_of_sstable() - skip remaining partitions in sstable
-    virtual future<stop_iteration> consume(partition_end&&) = 0;
-    // stop_iteration::yes -> full stop - skip remaining sstables
-    virtual future<stop_iteration> on_end_of_sstable() = 0;
-    // called at the very end
-    virtual future<> on_end_of_stream() = 0;
-};
 
 class consumer_wrapper {
 public:
